Move notes related logic to separate storage
