+++ AirCasting/CoreData/Storages/UIStorage.swift
@@ -23,7 +23,7 @@ final class CoreDataUIStorage: UIStorage {
         self.context = context
     }
     
-    /// All actions performed on CoreDataMeasurementStreamStorage must be performed
+    /// All actions performed on CoreDataUIStorage must be performed
     /// within a block passed to this methood.
     /// This ensures thread-safety by dispatching all calls to the queue owned by the NSManagedObjectContext.
     func accessStorage(_ task: @escaping(HiddenCoreDataUIStorage) -> Void) {
