+++ AppDelegate+Injection.swift
@@ -77,6 +77,7 @@ extension Resolver: ResolverRegistering {
         .implements(SessionUpdateable.self)
         .scope(.application)
         main.register { CoreDataMeasurementStreamStorage() as MeasurementStreamStorage }.scope(.cached)
+        main.register { DefaultSessionNotesStorage() as SessionNotesStorage }.scope(.cached)
         main.register { (_, _) -> UIStorage in
             let context = Resolver.resolve(PersistenceController.self).editContext
             return CoreDataUIStorage(context: context)
